Refactor jfr-agent-extension module into jfr-tools
