+++ settings.gradle.kts
@@ -9,7 +9,7 @@ pluginManagement {
 rootProject.name = "newrelic-jfr-core"
 
 include("jfr-daemon")
-include("jfr-agent-extension")
+include("jfr-tools")
 include("jfr-jlink")
 include("smoke-tests")
 include("jfr-mappers")
